+++ src/pycb/__init__.py
@@ -1,1 +1,2 @@
 from .couchbase import Couchbase, PycbException
+from .couchbase import PycbKeyNotFound, PycbKeyExists